pyproject.toml
[tool.poetry]
name = "date_assistant"
version = "0.8.0"
description = "🗓 Classes and functions with intuitive names for common dates operations"
authors = ["Jorge Alvarado <alvaradosegurajorge@gmail.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/jalvaradosegura/date-assistant"
repository = "https://github.com/jalvaradosegura/date-assistant"

[tool.poetry.dependencies]
python = "^3.8"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-cov = "^2.12.1"
flake8 = "^3.9.2"
black = "^21.7b0"
isort = "^5.9.3"
coveralls = "^3.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 79
ensure_newline_before_comments = true
